Feat: 마이페이지에서 이력서 검토 답변 보여주기 기능 추가

+++ core/src/main/java/com/goodjob/core/domain/payment/dto/request/PaymentRequestDto.java
@@ -10,5 +10,5 @@ public class PaymentRequestDto {
 
     String paymentKey;
 
-    Integer amount;
+    Long amount;
 }
